requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
asgiref==3.5.0
autopep8==1.5.4
blinker==1.4
brotli==1.0.9
brotlipy==0.7.0
ca-certificates==2022.10.11=haa95532_0
cachetools==4.1.1
certifi==2022.9.24
cffi==1.15.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
cssselect2==0.7.0
defusedxml==0.7.1
deflate==0.3.0
dj-database-url==1.0.0
django==3.2.16
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.2
drf-yasg==1.20.0
facebook-sdk==3.1.0
faker==4.1.0
#freetype==2.10.4
fpdf2==2.6.0
freetype-py==2.3.0
fonttools==4.38.0
future==0.18.2
google-api-core==1.22.2
google-api-python-client==2.39.0
google-auth==1.21.2
google-auth-httplib2==0.1.0
googleapis-common-protos==1.52.0
gunicorn==20.1.0
html5lib==1.1
httplib2==0.20.4
idna==2.9
inflection==0.5.1
image==1.5.33
itypes==1.2.0
jinja2==2.11.2
#jpeg==9e
#krb5==1.19.2
lerc==0.1.0
#libdeflate==1.8
#libpng==1.6.37
libpng-bins==0.0.3
#libpq==12.9
#libprotobuf==3.13.0.1
#libtiff==4.4.0
#libwebp==1.2.4
#libwebp-base==1.2.4
#lz4-c==1.9.3
lz4==4.0.2
markupsafe==1.1.1
#msvc-runtime==14.29.30133
numpy>=1.21.6
oauthlib==3.1.0
openssl==1.1.1s
packaging==20.4
pandas==1.3.5
pangocffi==0.11.0
pillow==9.3.0
pip==22.2.2
postgres==4.0
protobuf==3.13.0
psycopg2>=2.9.5
psycopg2-binary>=2.9.5
psycopg2-pool>=1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.21
pydyf==0.5.0
pyjwt==1.7.1
pylibjpeg-libjpeg==1.3.2
pyopenssl==22.0.0
pyparsing==2.4.7
pyphen==0.13.2
pysocks==1.7.1
#python==3.7.13
python-coreapi==2.3.3
python-coreschema==0.0.4
python-dateutil==2.8.1
python-dotenv==0.21.0
python-twitter==3.5
python-xz==0.4.0
pytz==2020.1
reportlab==3.6.12
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.6
ruamel.yaml>=0.16.10
ruamel.yaml.clib>=0.2.0
setuptools==65.5.0
six==1.15.0
#sqlite==3.39.3
sqlparse==0.3.1
text-unidecode==1.3
tifffile==2021.11.2
tinycss2==1.2.1
tk==0.1.0
toml==0.10.1
typing-extensions==4.1.1
uritemplate==4.1.1
urllib3==1.25.9
vc==2018.7.10
#vs2015_runtime==14.27.29016
weasyprint>=57.1
webencodings==0.5.1
webp>=0.1.4
wheel>=0.24.0
whitenoise>=6.2.0
win_inet_pton==1.1.0
wincertstore==0.2
#xz==5.2.6
#zlib==1.2.11
zlib-state>=0.1.5
zopfli>=0.2.2
zstd==1.5.0
